requirements.txt
github3.py>=0.1a1
